refactor(plugins,trafficpolicy): unify header modifier conversion
