+++ pkg/kgateway/extensions2/pluginutils/headers.go
@@ -1,12 +1,26 @@
 package pluginutils
 
 import (
+	"errors"
+	"fmt"
+	"sort"
+
 	mutation_rulesv3 "github.com/envoyproxy/go-control-plane/envoy/config/common/mutation_rules/v3"
 	envoycorev3 "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
+	"istio.io/istio/pkg/kube/krt"
+	"k8s.io/utils/ptr"
 	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
+
+	sharedv1alpha1 "github.com/kgateway-dev/kgateway/v2/api/v1alpha1/shared"
+	"github.com/kgateway-dev/kgateway/v2/pkg/krtcollections"
 )
 
-func ConvertMutations(filter *gwv1.HTTPHeaderFilter) []*mutation_rulesv3.HeaderMutation {
+var (
+	ErrUnsupportedRemoveHeaderMutation = errors.New("remove mutation cannot be converted to append action")
+	ErrUnknownHeaderMutation           = errors.New("unknown header mutation")
+)
+
+func ConvertMutations(filter *gwv1.HTTPHeaderFilter) (mutations []*mutation_rulesv3.HeaderMutation) {
 	if filter == nil {
 		return nil
 	}
@@ -14,8 +28,6 @@ func ConvertMutations(filter *gwv1.HTTPHeaderFilter) []*mutation_rulesv3.HeaderM
 	if len(filter.Add) == 0 && len(filter.Set) == 0 && len(filter.Remove) == 0 {
 		return nil
 	}
-
-	var mutations []*mutation_rulesv3.HeaderMutation
 
 	for _, h := range filter.Add {
 		mutations = append(mutations, &mutation_rulesv3.HeaderMutation{
@@ -52,5 +64,120 @@ func ConvertMutations(filter *gwv1.HTTPHeaderFilter) []*mutation_rulesv3.HeaderM
 			},
 		})
 	}
+
 	return mutations
+}
+
+func ConvertHeaderFilter(
+	krtctx krt.HandlerContext,
+	from krtcollections.From,
+	secrets *krtcollections.SecretIndex,
+	filter *sharedv1alpha1.HTTPHeaderFilter,
+) (*gwv1.HTTPHeaderFilter, error) {
+	if filter == nil {
+		return nil, nil
+	}
+
+	if len(filter.Add) == 0 && len(filter.Set) == 0 && len(filter.Remove) == 0 {
+		return nil, nil
+	}
+
+	gwFilter := &gwv1.HTTPHeaderFilter{}
+	for _, h := range filter.Add {
+		resolved, err := resolveHeader(krtctx, from, secrets, h)
+		if err != nil {
+			return nil, fmt.Errorf("failed to resolve header '%s' %w", ptr.Deref(h.Name, "<noname>"), err)
+		}
+		for _, r := range resolved {
+			gwFilter.Add = append(gwFilter.Add, r)
+		}
+	}
+
+	for _, h := range filter.Set {
+		resolved, err := resolveHeader(krtctx, from, secrets, h)
+		if err != nil {
+			return nil, fmt.Errorf("failed to resolve header '%s' %w", ptr.Deref(h.Name, "<noname>"), err)
+		}
+		for _, r := range resolved {
+			gwFilter.Add = append(gwFilter.Set, r)
+		}
+	}
+
+	gwFilter.Remove = filter.Remove
+
+	return gwFilter, nil
+}
+
+func ConvertMutationsToOptions(mutations []*mutation_rulesv3.HeaderMutation) (options []*envoycorev3.HeaderValueOption, err error) {
+	for _, m := range mutations {
+		switch a := m.Action.(type) {
+		case *mutation_rulesv3.HeaderMutation_Append:
+			options = append(options, a.Append)
+		case *mutation_rulesv3.HeaderMutation_Remove:
+			return nil, ErrUnsupportedRemoveHeaderMutation
+		default:
+			return nil, ErrUnknownHeaderMutation
+		}
+	}
+
+	return options, nil
+}
+
+// resolveHeader resolves an HTTPHeader entry into one or more gwv1.HTTPHeader pairs.
+// When both name and key are absent on a secretRef entry, all Secret data entries are returned.
+func resolveHeader(
+	krtctx krt.HandlerContext,
+	from krtcollections.From,
+	secrets *krtcollections.SecretIndex,
+	h sharedv1alpha1.HTTPHeader,
+) ([]gwv1.HTTPHeader, error) {
+	if h.Value != nil {
+		return []gwv1.HTTPHeader{{Name: *h.Name, Value: *h.Value}}, nil
+	}
+
+	ref := h.SecretRef
+	ns := gwv1.Namespace(from.Namespace)
+	if ref.Namespace != nil {
+		ns = *ref.Namespace
+	}
+	secretRef := gwv1.SecretObjectReference{
+		Name:      ref.Name,
+		Namespace: &ns,
+	}
+
+	secret, err := secrets.GetSecret(krtctx, from, secretRef)
+	if err != nil {
+		return nil, fmt.Errorf("secret %s/%s: %w", ns, ref.Name, err)
+	}
+
+	// Both name and key absent: inject every entry in the Secret as a header.
+	if ref.Key == nil && h.Name == nil {
+		pairs := make([]gwv1.HTTPHeader, 0, len(secret.Data))
+		for k, v := range secret.Data {
+			pairs = append(pairs, gwv1.HTTPHeader{Name: gwv1.HTTPHeaderName(k), Value: string(v)})
+		}
+		sort.Slice(pairs, func(i, j int) bool { return pairs[i].Name < pairs[j].Name })
+		return pairs, nil
+	}
+
+	// Determine which key to look up: explicit key, or fall back to header name.
+	// At this point at least one of ref.Key or h.Name is non-nil (both-nil handled above).
+	var key string
+	if ref.Key != nil {
+		key = *ref.Key
+	} else {
+		key = string(*h.Name)
+	}
+
+	data, ok := secret.Data[key]
+	if !ok {
+		return nil, fmt.Errorf("secret %s/%s does not contain key %q", ns, ref.Name, key)
+	}
+
+	// Determine header name: explicit name, or fall back to key.
+	headerName := gwv1.HTTPHeaderName(key)
+	if h.Name != nil {
+		headerName = *h.Name
+	}
+	return []gwv1.HTTPHeader{{Name: headerName, Value: string(data)}}, nil
 }
